renaming of workspace

+++ src/prime_gen.rs
@@ -1,8 +1,8 @@
 use std::io;
 use std::time::Instant;
 
-fn input() -> u64 {
-    //! Asks user for input and converts it to u64 if possible
+fn input() -> u128 {
+    //! Asks user for input and converts it to u128 if possible
     //! Loops until user inputs a valid number
     loop {
         let mut n:String = String::new();
@@ -12,7 +12,7 @@ fn input() -> u64 {
             .expect("Failed to read the line!");
 
 
-        let n:u64 = match n.trim().parse() {
+        let n:u128 = match n.trim().parse() {
             Ok(num) => num,
             Err(_) => {
                 println!("Please input a number (positive, non-zero preferably)!");
@@ -24,7 +24,7 @@ fn input() -> u64 {
     }
 }
 
-fn test_three(x:&u64) -> bool {
+fn test_three(x:&u128) -> bool {
     //! Tests if number is divisible by 3
     //! Return bool
     match x % 3 {
@@ -33,7 +33,7 @@ fn test_three(x:&u64) -> bool {
     }
 }
 
-fn test_five(x:&u64) -> bool {
+fn test_five(x:&u128) -> bool {
     //! Tests if number is divisible by 5
     //! Return bool
     match x % 5 {
@@ -42,7 +42,7 @@ fn test_five(x:&u64) -> bool {
     }
 }
 
-fn test_seven(x:&u64) -> bool {
+fn test_seven(x:&u128) -> bool {
     //! Tests if number is divisible by 7
     //! Return bool
     match x % 7 {
@@ -51,10 +51,10 @@ fn test_seven(x:&u64) -> bool {
     }
 }
 
-fn test_prime(x:u64, prime_list:Vec<u64>) -> Vec<u64> {
+fn test_prime(x:u128, prime_list:Vec<u128>) -> Vec<u128> {
     //! Tests all already generated numbers and if there's no match adds new number to the prime list
-    let mut counter:u64 = 0;
-    let mut prime_list:Vec<u64> = prime_list;
+    let mut counter:u128 = 0;
+    let mut prime_list:Vec<u128> = prime_list;
 
     for i in &prime_list {
 
@@ -76,14 +76,14 @@ fn test_prime(x:u64, prime_list:Vec<u64>) -> Vec<u64> {
     return prime_list;
 }
 
-fn primes(number:u64) -> Vec<u64> {
+fn primes(number:u128) -> Vec<u128> {
     //! Generates given number of primes
-    let mut prime_list:Vec<u64> = Vec::new();
+    let mut prime_list:Vec<u128> = Vec::new();
     
     match number {
         // Handles generatin prime list from size o to size 4 
         0..=4 => {
-            let preload:[u64; 4] = [2, 3, 5, 7];
+            let preload:[u128; 4] = [2, 3, 5, 7];
 
             if number == 0 {
                 return prime_list;
@@ -96,16 +96,16 @@ fn primes(number:u64) -> Vec<u64> {
         // All other numbers
         _ => {
 
-            let preload:[u64; 4] = [2, 3, 5, 7];
+            let preload:[u128; 4] = [2, 3, 5, 7];
             for i in preload { 
                 prime_list.push(i)
             }
 
-            let mut current:u64 = 9;
+            let mut current:u128 = 9;
 
             while number > prime_list.len().try_into().unwrap() {
 
-                let curry:u64 = current;
+                let curry:u128 = current;
 
                 if test_three(&curry) & test_five(&curry) & test_seven(&curry) {
                     prime_list = test_prime(curry, prime_list);
@@ -118,7 +118,7 @@ fn primes(number:u64) -> Vec<u64> {
     return prime_list;
 }
 
-fn printing(prime_list:&Vec<u64>) {
+fn printing(prime_list:&Vec<u128>) {
     //! Asks user if they want to print all of the results
     println!("Do you want to print the results?");
     //println!("Please input the number");
@@ -126,7 +126,7 @@ fn printing(prime_list:&Vec<u64>) {
     println!("[2] - No");
 
     loop {
-        let answer:u64 = input();
+        let answer:u128 = input();
 
         match answer {
             1 => {
@@ -153,12 +153,12 @@ fn printing(prime_list:&Vec<u64>) {
 
 }
 
-pub fn main_primes() -> Vec<u64> {
+pub fn main_primes() -> Vec<u128> {
     //! main function that can be called from python
     
     // Get the size of the list from user
     println!("How many primes to generate?");
-    let number:u64 = input();
+    let number:u128 = input();
 
     // Start generating timer
     let start = Instant::now();
